requirements.txt
torch
transformers
faiss-gpu
sentence-transformers
langchain
langchain_community
accelerate>=0.26.0
docx2txt